move entity events to events.entity subpackage
