+++ core/src/main/java/com/p1_7/abstractengine/managers/impl/EntityManager.java
@@ -7,11 +7,11 @@ import com.p1_7.abstractengine.core.AbstractProperty;
 import com.p1_7.abstractengine.core.Entity;
 import com.p1_7.abstractengine.core.EntityRepository;
 import com.p1_7.abstractengine.core.Tag;
-import com.p1_7.abstractengine.events.EntityActiveChangedEvent;
-import com.p1_7.abstractengine.events.EntityAddedEvent;
-import com.p1_7.abstractengine.events.EntityPropertyChangedEvent;
-import com.p1_7.abstractengine.events.EntityRemovedEvent;
-import com.p1_7.abstractengine.events.EntityTagChangedEvent;
+import com.p1_7.abstractengine.events.entity.EntityActiveChangedEvent;
+import com.p1_7.abstractengine.events.entity.EntityAddedEvent;
+import com.p1_7.abstractengine.events.entity.EntityPropertyChangedEvent;
+import com.p1_7.abstractengine.events.entity.EntityRemovedEvent;
+import com.p1_7.abstractengine.events.entity.EntityTagChangedEvent;
 import com.p1_7.abstractengine.events.Event;
 import com.p1_7.abstractengine.managers.base.AbstractManager;
